added ability to decide what parts of the bootstrap process are required
some ROMs for the X/2 require the use of custom hijacking methods in order
to run properly.  these can potentially be overwritten by DXB's boot
service integrity check.

this patch introduces the use of an ini-like configuration file that roms
can implement, located at /system/etc/DroidXBootstrap.cfg.  In the event
that such a file exists, the settings are read in from it and the
bootstrap then uses them to determine what it should and should not do.

The default is to do everything the bootstrapper would normally do should
this file not be present.